added developers API and moved code from GUI5
